Split integration test suite into non-invasive integration and invasive functional suite
This allows us to run the `CurlTransportTest` in Workflows to cover the class correctly.
