--- phpunit-integration.xml.dist
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-
-<phpunit backupGlobals="false"
-         backupStaticAttributes="false"
-         bootstrap="./tests/bootstrap.php"
-         colors="true"
-         convertErrorsToExceptions="true"
-         convertNoticesToExceptions="true"
-         convertWarningsToExceptions="true"
-         processIsolation="false"
-         stopOnFailure="false">
-
-    <filter>
-        <whitelist processUncoveredFilesFromWhitelist="true">
-            <directory suffix=".php">./src/</directory>
-        </whitelist>
-    </filter>
-
-    <testsuites>
-        <testsuite name="BigBlueButton test suit">
-            <directory>./tests/integration/</directory>
-        </testsuite>
-    </testsuites>
-
-</phpunit>
