+++ phpunit.xml.dist
@@ -18,11 +18,14 @@
     </filter>
 
     <testsuites>
-        <testsuite name="BigBlueButton test suite">
+        <testsuite name="BigBlueButton unit test suite">
             <directory>./tests/unit/</directory>
         </testsuite>
         <testsuite name="BigBlueButton integration test suite">
             <directory>./tests/integration/</directory>
+        </testsuite>
+        <testsuite name="BigBlueButton functional test suite">
+            <directory>./tests/functional/</directory>
         </testsuite>
     </testsuites>
