+++ tests/functional/AbstractBigBlueButtonFunctionalTest.php
@@ -16,8 +16,9 @@
  * You should have received a copy of the GNU Lesser General Public License along
  * with BigBlueButton; if not, see <http://www.gnu.org/licenses/>.
  */
-namespace BigBlueButton;
+namespace BigBlueButton\Tests\Functional;
 
+use BigBlueButton\BigBlueButton;
 use BigBlueButton\Http\Transport\TransportInterface;
 use BigBlueButton\Parameters\DeleteRecordingsParameters;
 use BigBlueButton\Parameters\EndMeetingParameters;
@@ -26,12 +27,13 @@ use BigBlueButton\Parameters\GetRecordingsParameters;
 use BigBlueButton\Parameters\IsMeetingRunningParameters;
 use BigBlueButton\Parameters\JoinMeetingParameters;
 use BigBlueButton\Parameters\PublishRecordingsParameters;
+use BigBlueButton\TestCase;
 
 /**
  * Class BigBlueButtonIntegrationTest
  * @package BigBlueButton
  */
-abstract class AbstractBigBlueButtonIntegrationTest extends TestCase
+abstract class AbstractBigBlueButtonFunctionalTest extends TestCase
 {
     /**
      * @var BigBlueButton
@@ -200,7 +202,7 @@ abstract class AbstractBigBlueButtonIntegrationTest extends TestCase
     /* Get Default Config XML */
 
     /**
-     * @group integration-legacy
+     * @group functional-legacy
      */
     public function testGetDefaultConfigXML()
     {
@@ -209,7 +211,7 @@ abstract class AbstractBigBlueButtonIntegrationTest extends TestCase
     }
 
     /**
-     * @group integration-legacy
+     * @group functional-legacy
      */
     public function testSetConfigXML()
     {
