+++ tests/functional/BigBlueButtonWithPsrHttpClientTransport.php
@@ -18,7 +18,7 @@ declare(strict_types=1);
  * You should have received a copy of the GNU Lesser General Public License
  * along with littleredbutton/bigbluebutton-api-php. If not, see <http://www.gnu.org/licenses/>.
  */
-namespace BigBlueButton;
+namespace BigBlueButton\Tests\Functional;
 
 use BigBlueButton\Http\Transport\Bridge\PsrHttpClient\PsrHttpClientTransport;
 use BigBlueButton\Http\Transport\TransportInterface;
@@ -26,7 +26,7 @@ use Nyholm\Psr7\Factory\Psr17Factory;
 use Symfony\Component\HttpClient\CurlHttpClient;
 use Symfony\Component\HttpClient\Psr18Client;
 
-final class BigBlueButtonWithPsrHttpClientTransport extends AbstractBigBlueButtonIntegrationTest
+final class BigBlueButtonWithPsrHttpClientTransport extends AbstractBigBlueButtonFunctionalTest
 {
     /**
      * {@inheritDoc}
