+++ tests/functional/BigBlueButtonWithSymfonyHttpClientTransportTest.php
@@ -18,18 +18,18 @@ declare(strict_types=1);
  * You should have received a copy of the GNU Lesser General Public License
  * along with littleredbutton/bigbluebutton-api-php. If not, see <http://www.gnu.org/licenses/>.
  */
-namespace BigBlueButton;
+namespace BigBlueButton\Tests\Functional;
 
-use BigBlueButton\Http\Transport\CurlTransport;
+use BigBlueButton\Http\Transport\Bridge\SymfonyHttpClient\SymfonyHttpClientTransport;
 use BigBlueButton\Http\Transport\TransportInterface;
 
-final class BigBlueButtonWithCurlTransportTest extends AbstractBigBlueButtonIntegrationTest
+final class BigBlueButtonWithSymfonyHttpClientTransportTest extends AbstractBigBlueButtonFunctionalTest
 {
     /**
      * {@inheritDoc}
      */
     protected static function createTransport(): TransportInterface
     {
-        return CurlTransport::createWithDefaultOptions();
+        return SymfonyHttpClientTransport::create();
     }
 }
